+++ packages/core/styles/src/packages/atomic/popover/nimbus-popover.css.ts
@@ -1,4 +1,4 @@
-import { style, styleVariants } from "@vanilla-extract/css";
+import { style } from "@vanilla-extract/css";
 import {
   createRainbowSprinkles,
   defineProperties as defineRainbowProperties,
@@ -18,7 +18,6 @@ const content = style({
   display: "flex",
   justifyContent: "center",
   alignItems: "center",
-  width: "17.5rem",
   minHeight: "2rem",
   filter: `drop-shadow(0px 0px 2px rgba(0, 0, 0, 0.2))`,
   boxSizing: "border-box",
@@ -27,84 +26,8 @@ const content = style({
   zIndex: varsThemeBase.zIndex[800],
 });
 
-const base = style({
-  "::after": {
-    content: " ",
-    position: "absolute",
-    borderWidth: "0.5rem",
-    borderStyle: "solid",
-  },
-});
-
-const content__arrow = styleVariants({
-  top: [
-    base,
-    {
-      "::after": {
-        top: "100%",
-        color: "currentcolor",
-        borderColor: `currentColor transparent transparent transparent`,
-        marginLeft: "-0.5rem",
-        left: "50%",
-      },
-    },
-  ],
-  bottom: [
-    base,
-    {
-      "::after": {
-        bottom: "100%",
-        borderColor: `transparent transparent currentColor transparent`,
-        marginLeft: "-0.5rem",
-        left: "50%",
-      },
-    },
-  ],
-  left: [
-    base,
-    {
-      "::after": {
-        bottom: "100%",
-        borderColor: `transparent transparent transparent currentColor`,
-        marginTop: "-0.5rem",
-        top: "50%",
-        left: "100%",
-      },
-    },
-  ],
-  right: [
-    base,
-    {
-      "::after": {
-        bottom: "100%",
-        borderColor: `transparent currentColor transparent transparent`,
-        marginTop: "-0.5rem",
-        top: "50%",
-        right: "100%",
-      },
-    },
-  ],
-});
-
-export const content__placement = styleVariants({
-  top: {},
-  right: {},
-  bottom: {},
-  left: {},
-  "top-start": {},
-  "top-end": {},
-  "right-start": { "::after": { marginTop: 0 } },
-  "right-end": { "::after": { marginTop: "-1rem" } },
-  "bottom-start": {},
-  "bottom-end": {},
-  "left-start": { "::after": { marginTop: 0 } },
-  "left-end": { "::after": { marginTop: "-1rem" } },
-});
-
 export const styles = {
   content,
-  content__arrow,
-  content__placement,
 };
 
 /* -------------------------------------------------------------------------------------------------
@@ -147,6 +70,7 @@ const defineProperties = defineRainbowProperties({
   defaultCondition: "xs",
   dynamicProperties: {
     width: true,
+    maxWidth: true,
     height: true,
   },
   staticProperties: {