+++ packages/core/styles/src/packages/atomic/tooltip/nimbus-tooltip.css.ts
@@ -1,4 +1,4 @@
-import { style, styleVariants } from "@vanilla-extract/css";
+import { style } from "@vanilla-extract/css";
 
 import { varsThemeBase } from "../../../themes";
 
@@ -13,62 +13,4 @@ export const content = style({
   borderRadius: varsThemeBase.shape.border.radius[1],
   transition: `opacity ${varsThemeBase.motion.speed.fast} ease`,
   zIndex: varsThemeBase.zIndex[800],
-});
-
-export const base = style({
-  "::after": {
-    content: " ",
-    position: "absolute",
-    borderWidth: varsThemeBase.shape.border.width[1],
-    borderStyle: "solid",
-  },
-});
-
-export const content__arrow = styleVariants({
-  top: [
-    base,
-    {
-      "::after": {
-        top: "100%",
-        borderColor: `${varsThemeBase.colors.neutral.textLow} transparent transparent transparent`,
-        marginLeft: "-0.25rem",
-        left: "50%",
-      },
-    },
-  ],
-  bottom: [
-    base,
-    {
-      "::after": {
-        bottom: "100%",
-        borderColor: `transparent transparent ${varsThemeBase.colors.neutral.textLow} transparent`,
-        marginLeft: "-0.25rem",
-        left: "50%",
-      },
-    },
-  ],
-  left: [
-    base,
-    {
-      "::after": {
-        bottom: "100%",
-        borderColor: `transparent transparent transparent ${varsThemeBase.colors.neutral.textLow}`,
-        marginTop: "-0.25rem",
-        top: "50%",
-        left: "100%",
-      },
-    },
-  ],
-  right: [
-    base,
-    {
-      "::after": {
-        bottom: "100%",
-        borderColor: `transparent ${varsThemeBase.colors.neutral.textLow} transparent transparent`,
-        marginTop: "-0.25rem",
-        top: "50%",
-        right: "100%",
-      },
-    },
-  ],
 });